+++ src/test/java/core/demo/app/integration/veiculo/VeiculoIntegrationTest.java
@@ -21,6 +21,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.jdbc.Sql;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -92,7 +93,7 @@ class VeiculoIntegrationTest {
         assertEquals(request.getModeloId(), createdOne.getModelo().getFipeId());
         assertEquals(request.getMarcaId(), createdOne.getMarca().getFipeId());
         assertEquals(request.getPrecoAnuncio(), createdOne.getPrecoAnuncio());
-        assertEquals(2_218_500, createdOne.getPrecoFipe());
+        assertEquals(BigInteger.valueOf(2_218_500), createdOne.getPrecoFipe());
         assertNotNull(createdOne.getCreatedAt());
 
         // 3. Read saved vehicle
