+++ src/test/java/core/demo/app/templates/VeiculoRequestTemplates.java
@@ -2,6 +2,8 @@ package core.demo.app.templates;
 
 import core.demo.app.interfaces.veiculo.VeiculoRequest;
 
+import java.math.BigInteger;
+
 public class VeiculoRequestTemplates {
 
     public static final String PLACA = "XYZ-9876";
@@ -17,7 +19,7 @@ public class VeiculoRequestTemplates {
                 .placa(PLACA)
                 .marcaId(21)
                 .modeloId(7754)
-                .precoAnuncio(10000000)
+                .precoAnuncio(BigInteger.valueOf(10000000))
                 .ano(2011);
     }
